REFACTORED Tooltip in Module into ModuleInfoTooltip
* Fixes #319
